Implement initial FormBuilder extension
Integration with [`ActionView`][actionview], [`form_with`][form_with],
[`form_for`][form_for], and Rails in general hinges on the
`ViewPartialFormBuilder::Engine` configuring
`ViewPartialFormBuilder::FormBuilder` as the host application's
[`default_form_builder`][default_form_builder] within an
[`ActiveSupport.on_load(:action_controller_base)` block][on_load].

According to the documentation:

> calls with `:action_controller_base` hooks will be called in the
> context of `ActionController::Base` (that means self will be an
> `ActionController::Base`).

Once configured as the default `<form>` builder,
`ViewPartialFormBuilder::FormBuilder` overrides various `<form>`-field-
builder instance methods (for example, `#label`, `#text_field`,
`#submit`) so that they first check for the existence of a matching
[`ActionView` partial template][partial_renderer].

When a partial template exists, render it. During rendering, the partial
is rendered either as a [`partial:`, or as a
`layout:`][partial_renderer], depending on if the call site invoked the
method with a [block][] or not.

The `ViewPartialFormBuilder::FormBuilder` will render the template,
making all of its arguments available to the template through the
`locals:` option. Within the view partial, the arguments will be
available:

* by name (for example, `method`, `options`)
* as part of the `arguments` array
* through the `local_assigns` hash

When a partial template does not exist, the
`ViewPartialFormBuilder::FormBuilder` falls back to rendering fields by
delegating to its ancestor's methods.

Additionally, This commit removes various Rails-generated files
configurations that are not necessary for `ActionView` integration.

[actionview]: https://guides.rubyonrails.org/action_view_overview.html
[form_with]: https://guides.rubyonrails.org/form_helpers.html#dealing-with-basic-forms
[form_for]: https://api.rubyonrails.org/v5.1/classes/ActionView/Helpers/FormHelper.html
[default_form_builder]: https://api.rubyonrails.org/v6.0/classes/ActionController/FormBuilder.html
[view_partial]: https://guides.rubyonrails.org/layouts_and_rendering.html#using-partials
[partial_renderer]: https://api.rubyonrails.org/classes/ActionView/PartialRenderer.html
[block]: https://ruby-doc.org/core-2.7.1/doc/syntax/methods_rdoc.html#label-Block+Argument
[on_load]: https://guides.rubyonrails.org/engines.html#how-does-it-work-questionmark

Testing
---

Tests that inherit from `FormBuilderTestCase` can declare `ActionView`
templates (with `#declare_template`) and `#render` them without needing
access to `Dummy::Application` or an `ActionController::Base` instance.

Additionally, tests that cover general integration, and test covering
the test hardness' `Dummy::Application` controller actions mixin the
`TemplateDeclarationHelpers` module for declaring `ActionView` partial
templates direclty in-line within test blocks.

[uniformity-view-path]: https://github.com/thoughtbot/uniformity/blob/b2d4706ec7c324a5399b47409966195ecf32d5a4/lib/uniformity/form_renderer.rb#L172-L181