Enable composing partials for the same field
Background
---

Prior to this commit, the `ViewPartialFormBuilder::FormBuilder` would
guard rendering view partials with a `#about_to_recurse_infinitely?`
guard. The guard was in place to prevent self-referential infinite
recursions like the following:

```html+erb
<%# app/views/form_builder/_label.html.erb %>

<%= form.label(
  *arguments,
  partial: "form_builder/label",
  **options,
  &block
) %>
```

To counteract that, after the first invocation, the `form` variable made
available to the `form_builder/label` partial was replaced with an
instance of [`ActionView::Helpers::FormBuilder`][form_builder], instead
of the `ViewPartialFormBuilder::FormBuilder` instance.

Problem
---

Passing a `partial:` key would be useful for layering partials on top of
one another. For instance, consider an administrative interface that
shared styles with a consumer facing site, but had additional bells and
whistles.

Solution
---

Extend the `ViewPartialFormBuilder::FormBuilder`'s check for
`#about_to_recurse_infinitely?` to allow for matching field names when a
`partial:` option is present.

Preventing infinite recursion is still of the utmost importance.
However, if a `partial:` option is present, check for it being a
[different `virtual_path`][actionview] than the current render, and
allow it when we're sure the next render cycle won't be
self-referential.

[form_builder]: https://api.rubyonrails.org/v6.0/classes/ActionView/Helpers/FormHelper.html
[actionview]: https://api.rubyonrails.org/classes/ActionView/Base.html