Improve test coverage of intended behavior
These changes more accurately exercise the _behavior_ of the helpers
they cover.

For example, prior to this commit, most `collection_`-prefixed helpers
were testing that the variables themselves were made available to the
partial.

While valuable, that doesn't accurately cover that the
`ViewPartialFormBuilder` constructs the resulting HTML to match the
behavior of the default `FormBuilder`.

Testing structure changes
---

In addition to writing better assertions, this commit moves the
`form_with` calls from in-test templates to the [Exercise Phase][] of
the test by passing the ERB template directly the the [`render` call as
the `inline:` option][render].

[Exercise Phase]: https://thoughtbot.com/blog/four-phase-test
[render]: https://api.rubyonrails.org/classes/ActionController/Renderer.html#method-i-render